cart

+++ core/apps/cart/views.py
@@ -1,11 +1,5 @@
-from django.shortcuts import render
-from django.views import View
-from .models import Cart,CartItem
-from ..books.models import Book
-# Create your views here.
-
-
-from django.shortcuts import get_object_or_404
+from django.shortcuts import render, get_object_or_404, redirect
+from django.http import Http404
 from django.views.generic import TemplateView
 from .models import Cart, CartItem
 from ..books.models import Book
@@ -16,25 +10,42 @@ class CartGenericView(TemplateView):
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         action = self.kwargs.get('action')
-        cart_item_id = self.kwargs.get('cart_item_id')  # для add это book_id, для remove — cart_item_id
+        cart_item_id = self.kwargs.get('cart_item_id')
+
+        # Проверка, что пользователь аутентифицирован
+        if not self.request.user.is_authenticated:
+            return redirect('login')  # Перенаправление на страницу входа
+
+        cart = Cart.objects.filter(user=self.request.user).first()
 
         if action == 'add':
             book = get_object_or_404(Book, id=cart_item_id)
-            cart, created = Cart.objects.get_or_create(
-                user=self.request.user,
-                defaults={'total_price': 0.00}
-            )
+            if not cart:
+                cart = Cart.objects.create(user=self.request.user)
             cart_item, created = CartItem.objects.get_or_create(cart=cart, book=book)
             cart_item.quantity += 1
             cart_item.save()
             cart.calculate_total_price()
+
         elif action == 'remove':
-            cart_item = get_object_or_404(CartItem, id=cart_item_id)
-            cart_item.delete()
-            cart = get_object_or_404(Cart, user=self.request.user)
+            if not cart:
+                raise Http404("Cart not found")
+
+            cart_item = CartItem.objects.filter(cart=cart, book_id=cart_item_id).first()
+            if not cart_item:
+                raise Http404("CartItem not found")
+
+            if cart_item.quantity > 1:
+                cart_item.quantity -= 1
+                cart_item.save()
+            else:
+                cart_item.delete()
+
             cart.calculate_total_price()
+
         else:
-            cart = None
+            if not cart:
+                cart = None
 
         context['cart'] = cart
         return context
